Fix hard-coded header sync

+++ cmd/headers/download/downloader.go
@@ -747,7 +747,7 @@ func (cs *ControlServerImpl) handleInboundMessage(ctx context.Context, inreq *pr
 func (cs *ControlServerImpl) sendHeaderRequest(ctx context.Context, req *headerdownload.HeaderRequest) []byte {
 	//log.Info(fmt.Sprintf("Sending header request {hash: %x, height: %d, length: %d}", req.Hash, req.Number, req.Length))
 	reqData := &eth.GetBlockHeadersData{
-		Amount:  uint64(req.Length),
+		Amount:  req.Length,
 		Reverse: req.Reverse,
 		Skip:    req.Skip,
 		Origin:  eth.HashOrNumber{Hash: req.Hash},
@@ -762,7 +762,7 @@ func (cs *ControlServerImpl) sendHeaderRequest(ctx context.Context, req *headerd
 	}
 	minBlock := req.Number
 	if !req.Reverse {
-		minBlock = req.Number + uint64(req.Length)*req.Skip
+		minBlock = req.Number + req.Length*req.Skip
 	}
 	outreq := proto_sentry.SendMessageByMinBlockRequest{
 		MinBlock: minBlock,
